create test file
