+++ .gitignore
@@ -1,3 +1,4 @@
 build
 .vscode
 build/
+.vscode/
