+++ CMakeLists.txt
@@ -2,4 +2,5 @@ cmake_minimum_required(VERSION 3.10)
 
 project(FindRoute)
 
-add_executable(findroute findroute.cpp)
+#add_executable(findroute findroute.cpp)
+add_executable(example findroute.cpp findroutetest.cpp)
